moved XSectionConfig aside
